Move app and component previews to common modules now that previews are supported in AS Narwhal. Leave shader previews where they are since they're more convenient where they are.
