--- app/src/androidMain/kotlin/com/alexrdclement/uiplayground/app/demo/control/Controls.android.kt
@@ -1,27 +1,0 @@
-package com.alexrdclement.uiplayground.app.demo.control
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.alexrdclement.uiplayground.app.preview.UiPlaygroundPreview
-import kotlinx.collections.immutable.persistentListOf
-
-@Preview
-@Composable
-private fun Preview() {
-    UiPlaygroundPreview {
-        Controls(
-            controls = persistentListOf(
-                Control.Slider(
-                    name = "Amount",
-                    value = 0.5f,
-                    onValueChange = {},
-                ),
-                Control.Slider(
-                    name = "Amount 2",
-                    value = 0.5f,
-                    onValueChange = {},
-                )
-            ),
-        )
-    }
-}
