--- app/src/androidMain/kotlin/com/alexrdclement/uiplayground/app/demo/shaders/ShaderScreen.android.kt
@@ -1,15 +1,0 @@
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.alexrdclement.uiplayground.app.demo.shaders.ShaderScreen
-import com.alexrdclement.uiplayground.app.preview.UiPlaygroundPreview
-
-@Preview
-@Composable
-private fun Preview() {
-    UiPlaygroundPreview {
-        ShaderScreen(
-            onNavigateBack = {},
-            onConfigureClick = {},
-        )
-    }
-}
