--- app/src/androidMain/kotlin/com/alexrdclement/uiplayground/app/demo/subject/DemoText.android.kt
@@ -1,13 +1,0 @@
-package com.alexrdclement.uiplayground.app.demo.subject
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.alexrdclement.uiplayground.app.preview.UiPlaygroundPreview
-
-@Preview
-@Composable
-private fun Preview() {
-    UiPlaygroundPreview {
-        DemoText()
-    }
-}
