+++ app/src/commonMain/kotlin/com/alexrdclement/uiplayground/app/demo/control/DropdownControl.kt
@@ -13,12 +13,16 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.BlurredEdgeTreatment
+import com.alexrdclement.uiplayground.app.preview.UiPlaygroundPreview
 import com.alexrdclement.uiplayground.components.Button
 import com.alexrdclement.uiplayground.components.DropdownMenu
 import com.alexrdclement.uiplayground.components.DropdownMenuItem
 import com.alexrdclement.uiplayground.components.Text
 import com.alexrdclement.uiplayground.components.VerticalDivider
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
+import kotlinx.collections.immutable.toPersistentList
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 fun <T> DropdownControl(
@@ -114,5 +118,59 @@ private fun <T> DropdownControlMenu(
                 }
             )
         }
+    }
+}
+
+@Preview
+@Composable
+private fun DropdownControlPreview() {
+    UiPlaygroundPreview {
+        var selectedIndex by remember { mutableStateOf(0) }
+        val control by remember {
+            mutableStateOf(
+                Control.Dropdown(
+                    name = "Edge treatment",
+                    values = listOf(
+                        BlurredEdgeTreatment.Rectangle,
+                        BlurredEdgeTreatment.Unbounded
+                    ).map {
+                        Control.Dropdown.DropdownItem(
+                            name = it.toString(),
+                            value = it
+                        )
+                    }.toPersistentList(),
+                    selectedIndex = selectedIndex,
+                    onValueChange = { selectedIndex = it }
+                )
+            )
+        }
+        DropdownControl(control = control)
+    }
+}
+
+@Preview
+@Composable
+private fun DropdownControlRowPreview() {
+    UiPlaygroundPreview {
+        var selectedIndex by remember { mutableStateOf(0) }
+        val control by remember {
+            mutableStateOf(
+                Control.Dropdown(
+                    name = "Edge treatment",
+                    values = listOf(
+                        BlurredEdgeTreatment.Rectangle,
+                        BlurredEdgeTreatment.Unbounded
+                    ).map {
+                        Control.Dropdown.DropdownItem(
+                            name = it.toString(),
+                            value = it
+                        )
+                    }.toPersistentList(),
+                    selectedIndex = selectedIndex,
+                    onValueChange = { selectedIndex = it }
+                )
+            )
+        }
+        DropdownControlRow(control = control)
     }
 }
