+++ app/src/commonMain/kotlin/com/alexrdclement/uiplayground/app/demo/control/SliderControl.kt
@@ -3,12 +3,18 @@ package com.alexrdclement.uiplayground.app.demo.control
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
+import com.alexrdclement.uiplayground.app.preview.UiPlaygroundPreview
 import com.alexrdclement.uiplayground.components.Slider
 import com.alexrdclement.uiplayground.components.Text
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 fun SliderControl(
@@ -28,5 +34,21 @@ fun SliderControl(
                 contentDescription = control.name
             }
         )
+    }
+}
+
+@Preview
+@Composable
+private fun Preview() {
+    UiPlaygroundPreview {
+        var value by remember { mutableStateOf(0.5f) }
+        val control = remember {
+            Control.Slider(
+                name = "Amount",
+                value = value,
+                onValueChange = { value = it },
+            )
+        }
+        SliderControl(control = control)
     }
 }
