+++ app/src/commonMain/kotlin/com/alexrdclement/uiplayground/app/demo/control/SubjectModifierBar.kt
@@ -5,17 +5,24 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import com.alexrdclement.uiplayground.app.demo.shaders.DemoModifier
 import com.alexrdclement.uiplayground.app.demo.subject.DemoSubject
+import com.alexrdclement.uiplayground.app.preview.UiPlaygroundPreview
+import com.alexrdclement.uiplayground.shaders.ColorSplitMode
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
 import kotlinx.collections.immutable.toPersistentList
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 fun SubjectModifierBar(
@@ -69,6 +76,45 @@ fun SubjectModifierBar(
             includeTitle = false,
             modifier = Modifier.semantics {
                 contentDescription = "Select modifier"
+            },
+        )
+    }
+}
+
+
+@Preview
+@Composable
+private fun Preview() {
+    UiPlaygroundPreview {
+        var demoSubject by remember { mutableStateOf(DemoSubject.Circle) }
+        val demoModifiers = remember {
+            mutableStateListOf<DemoModifier>(
+                DemoModifier.ColorSplit(
+                    xAmount = 0f,
+                    yAmount = 0f,
+                    colorMode = ColorSplitMode.RGB,
+                )
+            )
+        }
+        var demoModifierIndex by remember { mutableIntStateOf(0) }
+        val demoModifier by remember(demoModifiers, demoModifierIndex) {
+            derivedStateOf { demoModifiers[demoModifierIndex] }
+        }
+
+        var isSubjectDropdownExpanded by remember { mutableStateOf(true) }
+        var isModifierDropdownExpanded by remember { mutableStateOf(true) }
+
+        SubjectModifierBar(
+            demoSubject = demoSubject,
+            demoModifier = demoModifier,
+            demoModifiers = demoModifiers,
+            onSubjectSelected = {
+                demoSubject = it
+                isSubjectDropdownExpanded = false
+            },
+            onModifierSelected = {
+                demoModifierIndex = it
+                isModifierDropdownExpanded = false
             },
         )
     }
