--- components/src/androidMain/kotlin/com/alexrdclement/uiplayground/components/Checkbox.android.kt
@@ -1,28 +1,0 @@
-package com.alexrdclement.uiplayground.components
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.alexrdclement.uiplayground.components.preview.BoolPreviewParameterProvider
-import com.alexrdclement.uiplayground.theme.PlaygroundTheme
-
-@Preview
-@Composable
-private fun Preview(
-    @PreviewParameter(BoolPreviewParameterProvider::class) isChecked: Boolean,
-) {
-    PlaygroundTheme {
-        Surface {
-            var isChecked by remember { mutableStateOf(isChecked) }
-
-            Checkbox(
-                isChecked = isChecked,
-                onCheckedChange = { isChecked = !it },
-            )
-        }
-    }
-}
