+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/TopBar.kt
@@ -4,12 +4,15 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 fun TopBar(
@@ -38,6 +41,59 @@ fun TopBar(
         }
         Box(modifier = Modifier.padding(end = PlaygroundTheme.spacing.small)) {
             actions?.invoke()
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun Preview() {
+    PlaygroundTheme {
+        Surface {
+            TopBar(
+                title = {
+                    Text("Title", style = PlaygroundTheme.typography.headline)
+                }
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun NavButtonPreview() {
+    PlaygroundTheme {
+        Surface {
+            TopBar(
+                title = {
+                    Text("Title", style = PlaygroundTheme.typography.headline)
+                },
+                navButton = {
+                    BackNavigationButton(onClick = {})
+                }
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun ActionsPreview() {
+    PlaygroundTheme {
+        Surface {
+            TopBar(
+                title = {
+                    Text("Title", style = PlaygroundTheme.typography.headline)
+                },
+                actions = {
+                    Button(
+                        onClick = {},
+                        modifier = Modifier.size(48.dp),
+                    ) {
+                        Text("Action")
+                    }
+                }
+            )
         }
     }
 }
