+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/preview/BoolPreviewParameterProvider.kt
@@ -1,6 +1,6 @@
 package com.alexrdclement.uiplayground.components.preview
 
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameterProvider
 
 internal class BoolPreviewParameterProvider : PreviewParameterProvider<Boolean> {
     override val values = sequenceOf(true, false)
